Picking up changes from secdev/scapy#2347. SD can be folded into the SOMEIP code since they belong together.
